pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alertbox"
version = "0.1.0"
requires-python = ">=3.10"
description = "AlertBox Baltic: FastAPI app with schedulers and LLM fallback (Groq -> OpenAI)."
readme = "README.md"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "httpx>=0.27",
  "feedparser>=6.0",
  "beautifulsoup4>=4.12",
  "apscheduler>=3.10",
  "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.6",
  "black>=24.8",
  "isort>=5.13",
]

# ЯВНО: пакет только app/, чтобы не трогать deploy/, tests/
[tool.setuptools]
packages = ["app"]
include-package-data = true
